fix cursor

+++ src/header/driver/framebuffer.h
@@ -27,7 +27,7 @@
  * @param fg  Foreground / Character color
  * @param bg  Background color
  */
-void framebuffer_write(uint8_t row, uint8_t col, char c, uint8_t fg, uint8_t bg);
+void framebuffer_write(uint16_t row, uint16_t col, char c, uint8_t fg, uint8_t bg);
 
 /**
  * Set cursor to specified location. Row and column starts from 0
@@ -35,7 +35,7 @@ void framebuffer_write(uint8_t row, uint8_t col, char c, uint8_t fg, uint8_t bg)
  * @param r row
  * @param c column
 */
-void framebuffer_set_cursor(uint8_t r, uint8_t c);
+void framebuffer_set_cursor(uint16_t r, uint16_t c);
 
 /**
  * Set all cell in framebuffer character to 0x00 (empty character)
